fix(kb): bound serialized blob deletion
